# api/docs/swagger.yaml
definitions:
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
...
  user_service.CreateStudent:
    properties:
      branch_id:
        type: string
      course_count:
        type: integer
      full_name:
        type: string
      group_id:
        type: string
      paid_sum:
        type: number
      password:
        type: string
      phone:
        type: string
      total_sum:
        type: number
    type: object
  user_service.CreateSuperAdmin:
    properties:
      full_name:
        type: string
      ielts_score:
        type: number
      months_worked:
        type: integer
      password:
        type: string
      phone:
        type: string
      salary:
        type: number
    type: object
...
  user_service.UpdateStudent:
    properties:
      branch_id:
        type: string
      course_count:
        type: integer
      full_name:
        type: string
      group_id:
        type: string
      id:
        type: string
      paid_sum:
        type: number
      password:
        type: string
      phone:
        type: string
      total_sum:
        type: number
    type: object
  user_service.UpdateSuperAdmin:
    properties:
      full_name:
        type: string
      id:
        type: string
      ielts_score:
        type: number
      months_worked:
        type: integer
      password:
        type: string
      phone:
        type: string
      salary:
        type: number
    type: object
...
host: localhost:8082
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: CRM API
  version: "1.0"
paths:
...
  /api/v1/student:
    post:
      consumes:
      - application/json
      description: API for creating student
      parameters:
      - description: student
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/user_service.CreateStudent'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a student
      tags:
      - STUDENTS
    put:
      consumes:
      - application/json
      description: API for update a student
      parameters:
      - description: student
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/user_service.UpdateStudent'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Update a student
      tags:
      - STUDENTS
  /api/v1/student/{id}:
    delete:
      consumes:
      - application/json
      description: API for delete a student
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete a student
      tags:
      - STUDENTS
    get:
      consumes:
      - application/json
      description: API for getting a student
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get a student
      tags:
      - STUDENTS
  /api/v1/students:
    get:
      consumes:
      - application/json
      description: API for Get all students
      parameters:
      - description: search
        in: query
        name: search
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all students
      tags:
      - STUDENTS
  /api/v1/super-admin:
    post:
      consumes:
      - application/json
      description: API for creating super admin
      parameters:
      - description: super-admin
        in: body
        name: super-admin
        required: true
        schema:
          $ref: '#/definitions/user_service.CreateSuperAdmin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Create an super admin
      tags:
      - SUPER-ADMINS
    put:
      consumes:
      - application/json
      description: API for update super admin
      parameters:
      - description: super-admin
        in: body
        name: super-admin
        required: true
        schema:
          $ref: '#/definitions/user_service.UpdateSuperAdmin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Update super admin
      tags:
      - SUPER-ADMINS
  /api/v1/super-admin/{id}:
    delete:
      consumes:
      - application/json
      description: API for delete a super admin
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete super admin
      tags:
      - SUPER-ADMINS
    get:
      consumes:
      - application/json
      description: API for getting superadmin
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get an super admin
      tags:
      - SUPER-ADMINS
  /api/v1/super-admins:
    get:
      consumes:
      - application/json
      description: API for Get all super admins
      parameters:
      - description: search
        in: query
        name: search
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all super admins
      tags:
      - SUPER-ADMINS
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"